% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_activity_log.R
\docType{class}
\name{ArrayActivityLog}
\alias{ArrayActivityLog}
\title{ArrayActivityLog}
\format{
An \code{R6Class} generator object
}
\description{
ArrayActivityLog Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event_at}}{character [optional]}

\item{\code{action}}{\link{ActivityEventType} [optional]}

\item{\code{username}}{character [optional]}

\item{\code{bytes_sent}}{integer [optional]}

\item{\code{bytes_received}}{integer [optional]}

\item{\code{array_task_id}}{character [optional]}

\item{\code{query_ranges}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ArrayActivityLog$new()}}
\item \href{#method-toJSON}{\code{ArrayActivityLog$toJSON()}}
\item \href{#method-fromJSON}{\code{ArrayActivityLog$fromJSON()}}
\item \href{#method-toJSONString}{\code{ArrayActivityLog$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ArrayActivityLog$fromJSONString()}}
\item \href{#method-clone}{\code{ArrayActivityLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$new(
  event_at = NULL,
  action = NULL,
  username = NULL,
  bytes_sent = NULL,
  bytes_received = NULL,
  array_task_id = NULL,
  query_ranges = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$fromJSON(ArrayActivityLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$fromJSONString(ArrayActivityLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActivityLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
