% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_api.R
\docType{class}
\name{ArrayApi}
\alias{ArrayApi}
\title{Array operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Array
}
\section{Methods}{

\describe{
\strong{ ArrayActivityLog } \emph{  }
get array activity logs

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } start integer
\item \emph{ @param } end integer
\item \emph{ @param } event.types character
\item \emph{ @param } task.id character
\item \emph{ @returnType } list( \link{ArrayActivityLog} ) \cr


\item status code : 200 | log of array activity

\item return type : array[ArrayActivityLog] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserOwnedGet } \emph{  }
Fetch a list of all arrays that are owned directly by user or user&#39;s organizations

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @param } search character
\item \emph{ @param } namespace character
\item \emph{ @param } orderby character
\item \emph{ @param } permissions character
\item \emph{ @param } tag list( character )
\item \emph{ @returnType } \link{ArrayBrowserData} \cr


\item status code : 200 | Array of array info that are owned directly by user or user&#39;s organizations

\item return type : ArrayBrowserData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserOwnedSidebarGet } \emph{  }
Fetch a sidebar for arrays that are owned directly by user or user&#39;s organizations

\itemize{
\item \emph{ @returnType } \link{ArrayBrowserSidebar} \cr


\item status code : 200 | Array of array info that are owned directly by user or user&#39;s organizations

\item return type : ArrayBrowserSidebar 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserPublicGet } \emph{  }
Fetch a list of all arrays that have been shared publically

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @param } search character
\item \emph{ @param } namespace character
\item \emph{ @param } orderby character
\item \emph{ @param } permissions character
\item \emph{ @param } tag list( character )
\item \emph{ @returnType } \link{ArrayBrowserData} \cr


\item status code : 200 | Array of array info that has been shared publically

\item return type : ArrayBrowserData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserPublicSidebarGet } \emph{  }
Fetch a sidebar of all arrays that have been shared publically

\itemize{
\item \emph{ @returnType } \link{ArrayBrowserSidebar} \cr


\item status code : 200 | Array of array info that has been shared publically

\item return type : ArrayBrowserSidebar 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserSharedGet } \emph{  }
Fetch a list of all arrays that have been shared with the user

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @param } search character
\item \emph{ @param } namespace character
\item \emph{ @param } orderby character
\item \emph{ @param } permissions character
\item \emph{ @param } tag list( character )
\item \emph{ @returnType } \link{ArrayBrowserData} \cr


\item status code : 200 | Array of array info that has been shared with the user

\item return type : ArrayBrowserData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ArraysBrowserSharedSidebarGet } \emph{  }
Fetch a list of all arrays that have been shared with the user

\itemize{
\item \emph{ @returnType } \link{ArrayBrowserSidebar} \cr


\item status code : 200 | Array of array info that has been shared with the user

\item return type : ArrayBrowserSidebar 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ CreateArray } \emph{  }
create a array schema at a specified URI registered to a group/project

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } content.type character
\item \emph{ @param } array.schema \link{ArraySchema}


\item status code : 204 | schema created successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteArray } \emph{  }
delete a array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } content.type character


\item status code : 204 | delete array successful


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeregisterArray } \emph{  }
deregister a array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character


\item status code : 204 | deregistered array successful


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetAllArrayMetadata } \emph{  }
get all array metadata user has access to

\itemize{
\item \emph{ @param } public.share character
\item \emph{ @returnType } list( \link{ArrayInfo} ) \cr


\item status code : 200 | array metadata for all arrays user has access to

\item return type : array[ArrayInfo] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArray } \emph{  }
get an ArraySchema using a url encoded uri

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } content.type character
\item \emph{ @returnType } \link{ArraySchema} \cr


\item status code : 200 | get ArraySchema

\item return type : ArraySchema 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArrayMaxBufferSizes } \emph{  }
get the max buffer sizes of an array for a subarray

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } subarray character
\item \emph{ @param } content.type character
\item \emph{ @param } x.payer character
\item \emph{ @returnType } \link{MaxBufferSizes} \cr


\item status code : 200 | get the max buffer sizes of an array for a subarray

\item return type : MaxBufferSizes 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArrayMetaDataJson } \emph{  }
get metadata from the array in json format

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } length integer


\item status code : 200 | get array metadata

\item return type : object 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArrayMetadata } \emph{  }
get metadata on an array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @returnType } \link{ArrayInfo} \cr


\item status code : 200 | array metadata for an array

\item return type : ArrayInfo 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArrayNonEmptyDomain } \emph{  }
get the non empty domain of an array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } content.type character
\item \emph{ @param } x.payer character
\item \emph{ @returnType } \link{NonEmptyDomain} \cr


\item status code : 200 | get the non empty domain of an array

\item return type : NonEmptyDomain 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArraySampleData } \emph{  }
get an sample set of data from the array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } samples numeric
\item \emph{ @returnType } \link{ArraySample} \cr


\item status code : 200 | get array sample data

\item return type : ArraySample 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArraySharingPolicies } \emph{  }
Get all sharing details of the array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @returnType } list( \link{ArraySharing} ) \cr


\item status code : 200 | List of all specific sharing policies

\item return type : array[ArraySharing] 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | Array does not exist or user does not have permissions to view array sharing policies


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArraysInNamespace } \emph{  }
get metadata on all arrays in a namespace

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @returnType } list( \link{ArrayInfo} ) \cr


\item status code : 200 | array metadata for all arrays in a namespace

\item return type : array[ArrayInfo] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetLastAccessedArrays } \emph{  }


\itemize{
\item \emph{ @returnType } list( \link{LastAccessedArray} ) \cr


\item status code : 200 | gets last accessed arrays

\item return type : array[LastAccessedArray] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ RegisterArray } \emph{  }
register an array at a specified URI registered to the given namespace

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } array.metadata \link{ArrayInfoUpdate}


\item status code : 204 | schema registered successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ShareArray } \emph{  }
Share an array with a user

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } array.sharing \link{ArraySharing}


\item status code : 204 | Array shared successfully


\item response headers :

\tabular{ll}{
}
\item status code : 404 | Array does not exist or user does not have permissions to share array


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ UpdateArrayMetadata } \emph{  }
update metadata on an array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } array.metadata \link{ArrayInfoUpdate}


\item status code : 204 | array metadata updated successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  ArrayActivityLog  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.start <- 56 # integer | Start time of window of fetch logs, unix epoch in seconds (default: seven days ago)
var.end <- 56 # integer | End time of window of fetch logs, unix epoch in seconds (default: current utc timestamp)
var.event.types <- 'event.types_example' # character | Event values can be one or more of the following read, write, create, delete, register, deregister, comma separated
var.task.id <- 'task.id_example' # character | Array task id To filter activity to

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArrayActivityLog(var.namespace, var.array, start=var.start, end=var.end, event.types=var.event.types, task.id=var.task.id)


####################  ArraysBrowserOwnedGet  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit
var.search <- 'search_example' # character | search string that will look at name, namespace or description fields
var.namespace <- 'namespace_example' # character | namespace
var.orderby <- 'orderby_example' # character | sort by which field valid values include last_accessed, size, name
var.permissions <- 'permissions_example' # character | permissions valid values include read, read_write, write, admin
var.tag <- ['tag_example'] # array[character] | tag to search for, more than one can be included

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserOwnedGet(page=var.page, per.page=var.per.page, search=var.search, namespace=var.namespace, orderby=var.orderby, permissions=var.permissions, tag=var.tag)


####################  ArraysBrowserOwnedSidebarGet  ####################

library(tiledbcloud)

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserOwnedSidebarGet()


####################  ArraysBrowserPublicGet  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit
var.search <- 'search_example' # character | search string that will look at name, namespace or description fields
var.namespace <- 'namespace_example' # character | namespace
var.orderby <- 'orderby_example' # character | sort by which field valid values include last_accessed, size, name
var.permissions <- 'permissions_example' # character | permissions valid values include read, read_write, write, admin
var.tag <- ['tag_example'] # array[character] | tag to search for, more than one can be included

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserPublicGet(page=var.page, per.page=var.per.page, search=var.search, namespace=var.namespace, orderby=var.orderby, permissions=var.permissions, tag=var.tag)


####################  ArraysBrowserPublicSidebarGet  ####################

library(tiledbcloud)

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserPublicSidebarGet()


####################  ArraysBrowserSharedGet  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit
var.search <- 'search_example' # character | search string that will look at name, namespace or description fields
var.namespace <- 'namespace_example' # character | namespace
var.orderby <- 'orderby_example' # character | sort by which field valid values include last_accessed, size, name
var.permissions <- 'permissions_example' # character | permissions valid values include read, read_write, write, admin
var.tag <- ['tag_example'] # array[character] | tag to search for, more than one can be included

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserSharedGet(page=var.page, per.page=var.per.page, search=var.search, namespace=var.namespace, orderby=var.orderby, permissions=var.permissions, tag=var.tag)


####################  ArraysBrowserSharedSidebarGet  ####################

library(tiledbcloud)

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ArraysBrowserSharedSidebarGet()


####################  CreateArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.content.type <- 'application/json' # character | Content Type of input and return mime
var.array.schema <- ArraySchema$new() # ArraySchema | ArraySchema being created

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$CreateArray(var.namespace, var.array, var.content.type, var.array.schema)


####################  DeleteArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.content.type <- 'application/json' # character | Content Type of input and return mime

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteArray(var.namespace, var.array, var.content.type)


####################  DeregisterArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeregisterArray(var.namespace, var.array)


####################  GetAllArrayMetadata  ####################

library(tiledbcloud)
var.public.share <- 'public.share_example' # character | Public share values can be one of exclude, only

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetAllArrayMetadata(public.share=var.public.share)


####################  GetArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.content.type <- 'application/json' # character | Content Type of input and return mime

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArray(var.namespace, var.array, var.content.type)


####################  GetArrayMaxBufferSizes  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.subarray <- 'subarray_example' # character | CSV string of subarray to get max buffer sizes for
var.content.type <- 'application/json' # character | Content Type of input and return mime
var.x.payer <- 'x.payer_example' # character | Name of organization or user who should be charged for this request

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArrayMaxBufferSizes(var.namespace, var.array, var.subarray, var.content.type, x.payer=var.x.payer)


####################  GetArrayMetaDataJson  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.length <- 56 # integer | (optional) limit character length of returned values

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArrayMetaDataJson(var.namespace, var.array, length=var.length)


####################  GetArrayMetadata  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArrayMetadata(var.namespace, var.array)


####################  GetArrayNonEmptyDomain  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.content.type <- 'application/json' # character | Content Type of input and return mime
var.x.payer <- 'x.payer_example' # character | Name of organization or user who should be charged for this request

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArrayNonEmptyDomain(var.namespace, var.array, var.content.type, x.payer=var.x.payer)


####################  GetArraySampleData  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.samples <- 5.0 # numeric | Number of sample results to return

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArraySampleData(var.namespace, var.array, samples=var.samples)


####################  GetArraySharingPolicies  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArraySharingPolicies(var.namespace, var.array)


####################  GetArraysInNamespace  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArraysInNamespace(var.namespace)


####################  GetLastAccessedArrays  ####################

library(tiledbcloud)

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetLastAccessedArrays()


####################  RegisterArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.array.metadata <- ArrayInfoUpdate$new() # ArrayInfoUpdate | metadata associated with array

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$RegisterArray(var.namespace, var.array, var.array.metadata)


####################  ShareArray  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.array.sharing <- ArraySharing$new() # ArraySharing | Namespace and list of permissions to share with. An empty list of permissions will remove the namespace, if permissions already exist they will be deleted then new ones added. In the event of a failure, the new polcies will be rolled back to prevent partial policies, and its likely the array will not be shared with the namespace at all

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ShareArray(var.namespace, var.array, var.array.sharing)


####################  UpdateArrayMetadata  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.array.metadata <- ArrayInfoUpdate$new() # ArrayInfoUpdate | array metadata to update

api.instance <- ArrayApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$UpdateArrayMetadata(var.namespace, var.array, var.array.metadata)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ArrayApi$new()}}
\item \href{#method-ArrayActivityLog}{\code{ArrayApi$ArrayActivityLog()}}
\item \href{#method-ArrayActivityLogWithHttpInfo}{\code{ArrayApi$ArrayActivityLogWithHttpInfo()}}
\item \href{#method-ArraysBrowserOwnedGet}{\code{ArrayApi$ArraysBrowserOwnedGet()}}
\item \href{#method-ArraysBrowserOwnedGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserOwnedGetWithHttpInfo()}}
\item \href{#method-ArraysBrowserOwnedSidebarGet}{\code{ArrayApi$ArraysBrowserOwnedSidebarGet()}}
\item \href{#method-ArraysBrowserOwnedSidebarGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserOwnedSidebarGetWithHttpInfo()}}
\item \href{#method-ArraysBrowserPublicGet}{\code{ArrayApi$ArraysBrowserPublicGet()}}
\item \href{#method-ArraysBrowserPublicGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserPublicGetWithHttpInfo()}}
\item \href{#method-ArraysBrowserPublicSidebarGet}{\code{ArrayApi$ArraysBrowserPublicSidebarGet()}}
\item \href{#method-ArraysBrowserPublicSidebarGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserPublicSidebarGetWithHttpInfo()}}
\item \href{#method-ArraysBrowserSharedGet}{\code{ArrayApi$ArraysBrowserSharedGet()}}
\item \href{#method-ArraysBrowserSharedGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserSharedGetWithHttpInfo()}}
\item \href{#method-ArraysBrowserSharedSidebarGet}{\code{ArrayApi$ArraysBrowserSharedSidebarGet()}}
\item \href{#method-ArraysBrowserSharedSidebarGetWithHttpInfo}{\code{ArrayApi$ArraysBrowserSharedSidebarGetWithHttpInfo()}}
\item \href{#method-CreateArray}{\code{ArrayApi$CreateArray()}}
\item \href{#method-CreateArrayWithHttpInfo}{\code{ArrayApi$CreateArrayWithHttpInfo()}}
\item \href{#method-DeleteArray}{\code{ArrayApi$DeleteArray()}}
\item \href{#method-DeleteArrayWithHttpInfo}{\code{ArrayApi$DeleteArrayWithHttpInfo()}}
\item \href{#method-DeregisterArray}{\code{ArrayApi$DeregisterArray()}}
\item \href{#method-DeregisterArrayWithHttpInfo}{\code{ArrayApi$DeregisterArrayWithHttpInfo()}}
\item \href{#method-GetAllArrayMetadata}{\code{ArrayApi$GetAllArrayMetadata()}}
\item \href{#method-GetAllArrayMetadataWithHttpInfo}{\code{ArrayApi$GetAllArrayMetadataWithHttpInfo()}}
\item \href{#method-GetArray}{\code{ArrayApi$GetArray()}}
\item \href{#method-GetArrayWithHttpInfo}{\code{ArrayApi$GetArrayWithHttpInfo()}}
\item \href{#method-GetArrayMaxBufferSizes}{\code{ArrayApi$GetArrayMaxBufferSizes()}}
\item \href{#method-GetArrayMaxBufferSizesWithHttpInfo}{\code{ArrayApi$GetArrayMaxBufferSizesWithHttpInfo()}}
\item \href{#method-GetArrayMetaDataJson}{\code{ArrayApi$GetArrayMetaDataJson()}}
\item \href{#method-GetArrayMetaDataJsonWithHttpInfo}{\code{ArrayApi$GetArrayMetaDataJsonWithHttpInfo()}}
\item \href{#method-GetArrayMetadata}{\code{ArrayApi$GetArrayMetadata()}}
\item \href{#method-GetArrayMetadataWithHttpInfo}{\code{ArrayApi$GetArrayMetadataWithHttpInfo()}}
\item \href{#method-GetArrayNonEmptyDomain}{\code{ArrayApi$GetArrayNonEmptyDomain()}}
\item \href{#method-GetArrayNonEmptyDomainWithHttpInfo}{\code{ArrayApi$GetArrayNonEmptyDomainWithHttpInfo()}}
\item \href{#method-GetArraySampleData}{\code{ArrayApi$GetArraySampleData()}}
\item \href{#method-GetArraySampleDataWithHttpInfo}{\code{ArrayApi$GetArraySampleDataWithHttpInfo()}}
\item \href{#method-GetArraySharingPolicies}{\code{ArrayApi$GetArraySharingPolicies()}}
\item \href{#method-GetArraySharingPoliciesWithHttpInfo}{\code{ArrayApi$GetArraySharingPoliciesWithHttpInfo()}}
\item \href{#method-GetArraysInNamespace}{\code{ArrayApi$GetArraysInNamespace()}}
\item \href{#method-GetArraysInNamespaceWithHttpInfo}{\code{ArrayApi$GetArraysInNamespaceWithHttpInfo()}}
\item \href{#method-GetLastAccessedArrays}{\code{ArrayApi$GetLastAccessedArrays()}}
\item \href{#method-GetLastAccessedArraysWithHttpInfo}{\code{ArrayApi$GetLastAccessedArraysWithHttpInfo()}}
\item \href{#method-RegisterArray}{\code{ArrayApi$RegisterArray()}}
\item \href{#method-RegisterArrayWithHttpInfo}{\code{ArrayApi$RegisterArrayWithHttpInfo()}}
\item \href{#method-ShareArray}{\code{ArrayApi$ShareArray()}}
\item \href{#method-ShareArrayWithHttpInfo}{\code{ArrayApi$ShareArrayWithHttpInfo()}}
\item \href{#method-UpdateArrayMetadata}{\code{ArrayApi$UpdateArrayMetadata()}}
\item \href{#method-UpdateArrayMetadataWithHttpInfo}{\code{ArrayApi$UpdateArrayMetadataWithHttpInfo()}}
\item \href{#method-clone}{\code{ArrayApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActivityLog"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActivityLog}{}}}
\subsection{Method \code{ArrayActivityLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArrayActivityLog(
  namespace,
  array,
  start = NULL,
  end = NULL,
  event.types = NULL,
  task.id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActivityLogWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActivityLogWithHttpInfo}{}}}
\subsection{Method \code{ArrayActivityLogWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArrayActivityLogWithHttpInfo(
  namespace,
  array,
  start = NULL,
  end = NULL,
  event.types = NULL,
  task.id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserOwnedGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserOwnedGet}{}}}
\subsection{Method \code{ArraysBrowserOwnedGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserOwnedGet(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserOwnedGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserOwnedGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserOwnedGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserOwnedGetWithHttpInfo(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserOwnedSidebarGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserOwnedSidebarGet}{}}}
\subsection{Method \code{ArraysBrowserOwnedSidebarGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserOwnedSidebarGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserOwnedSidebarGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserOwnedSidebarGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserOwnedSidebarGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserOwnedSidebarGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserPublicGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserPublicGet}{}}}
\subsection{Method \code{ArraysBrowserPublicGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserPublicGet(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserPublicGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserPublicGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserPublicGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserPublicGetWithHttpInfo(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserPublicSidebarGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserPublicSidebarGet}{}}}
\subsection{Method \code{ArraysBrowserPublicSidebarGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserPublicSidebarGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserPublicSidebarGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserPublicSidebarGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserPublicSidebarGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserPublicSidebarGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserSharedGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserSharedGet}{}}}
\subsection{Method \code{ArraysBrowserSharedGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserSharedGet(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserSharedGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserSharedGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserSharedGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserSharedGetWithHttpInfo(
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserSharedSidebarGet"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserSharedSidebarGet}{}}}
\subsection{Method \code{ArraysBrowserSharedSidebarGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserSharedSidebarGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraysBrowserSharedSidebarGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ArraysBrowserSharedSidebarGetWithHttpInfo}{}}}
\subsection{Method \code{ArraysBrowserSharedSidebarGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ArraysBrowserSharedSidebarGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateArray"></a>}}
\if{latex}{\out{\hypertarget{method-CreateArray}{}}}
\subsection{Method \code{CreateArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$CreateArray(namespace, array, content.type, array.schema, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CreateArrayWithHttpInfo}{}}}
\subsection{Method \code{CreateArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$CreateArrayWithHttpInfo(
  namespace,
  array,
  content.type,
  array.schema,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteArray"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteArray}{}}}
\subsection{Method \code{DeleteArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$DeleteArray(namespace, array, content.type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteArrayWithHttpInfo}{}}}
\subsection{Method \code{DeleteArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$DeleteArrayWithHttpInfo(namespace, array, content.type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeregisterArray"></a>}}
\if{latex}{\out{\hypertarget{method-DeregisterArray}{}}}
\subsection{Method \code{DeregisterArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$DeregisterArray(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeregisterArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeregisterArrayWithHttpInfo}{}}}
\subsection{Method \code{DeregisterArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$DeregisterArrayWithHttpInfo(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAllArrayMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-GetAllArrayMetadata}{}}}
\subsection{Method \code{GetAllArrayMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetAllArrayMetadata(public.share = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAllArrayMetadataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetAllArrayMetadataWithHttpInfo}{}}}
\subsection{Method \code{GetAllArrayMetadataWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetAllArrayMetadataWithHttpInfo(public.share = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArray"></a>}}
\if{latex}{\out{\hypertarget{method-GetArray}{}}}
\subsection{Method \code{GetArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArray(namespace, array, content.type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayWithHttpInfo}{}}}
\subsection{Method \code{GetArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayWithHttpInfo(namespace, array, content.type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMaxBufferSizes"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMaxBufferSizes}{}}}
\subsection{Method \code{GetArrayMaxBufferSizes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMaxBufferSizes(
  namespace,
  array,
  subarray,
  content.type,
  x.payer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMaxBufferSizesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMaxBufferSizesWithHttpInfo}{}}}
\subsection{Method \code{GetArrayMaxBufferSizesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMaxBufferSizesWithHttpInfo(
  namespace,
  array,
  subarray,
  content.type,
  x.payer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMetaDataJson"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMetaDataJson}{}}}
\subsection{Method \code{GetArrayMetaDataJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMetaDataJson(namespace, array, length = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMetaDataJsonWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMetaDataJsonWithHttpInfo}{}}}
\subsection{Method \code{GetArrayMetaDataJsonWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMetaDataJsonWithHttpInfo(namespace, array, length = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMetadata}{}}}
\subsection{Method \code{GetArrayMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMetadata(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayMetadataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayMetadataWithHttpInfo}{}}}
\subsection{Method \code{GetArrayMetadataWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayMetadataWithHttpInfo(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayNonEmptyDomain"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayNonEmptyDomain}{}}}
\subsection{Method \code{GetArrayNonEmptyDomain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayNonEmptyDomain(
  namespace,
  array,
  content.type,
  x.payer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayNonEmptyDomainWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayNonEmptyDomainWithHttpInfo}{}}}
\subsection{Method \code{GetArrayNonEmptyDomainWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArrayNonEmptyDomainWithHttpInfo(
  namespace,
  array,
  content.type,
  x.payer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraySampleData"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraySampleData}{}}}
\subsection{Method \code{GetArraySampleData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraySampleData(namespace, array, samples = 5, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraySampleDataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraySampleDataWithHttpInfo}{}}}
\subsection{Method \code{GetArraySampleDataWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraySampleDataWithHttpInfo(namespace, array, samples = 5, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraySharingPolicies"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraySharingPolicies}{}}}
\subsection{Method \code{GetArraySharingPolicies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraySharingPolicies(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraySharingPoliciesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraySharingPoliciesWithHttpInfo}{}}}
\subsection{Method \code{GetArraySharingPoliciesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraySharingPoliciesWithHttpInfo(namespace, array, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraysInNamespace"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraysInNamespace}{}}}
\subsection{Method \code{GetArraysInNamespace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraysInNamespace(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArraysInNamespaceWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArraysInNamespaceWithHttpInfo}{}}}
\subsection{Method \code{GetArraysInNamespaceWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetArraysInNamespaceWithHttpInfo(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLastAccessedArrays"></a>}}
\if{latex}{\out{\hypertarget{method-GetLastAccessedArrays}{}}}
\subsection{Method \code{GetLastAccessedArrays()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetLastAccessedArrays(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetLastAccessedArraysWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetLastAccessedArraysWithHttpInfo}{}}}
\subsection{Method \code{GetLastAccessedArraysWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$GetLastAccessedArraysWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegisterArray"></a>}}
\if{latex}{\out{\hypertarget{method-RegisterArray}{}}}
\subsection{Method \code{RegisterArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$RegisterArray(namespace, array, array.metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegisterArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RegisterArrayWithHttpInfo}{}}}
\subsection{Method \code{RegisterArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$RegisterArrayWithHttpInfo(namespace, array, array.metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShareArray"></a>}}
\if{latex}{\out{\hypertarget{method-ShareArray}{}}}
\subsection{Method \code{ShareArray()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ShareArray(namespace, array, array.sharing, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShareArrayWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ShareArrayWithHttpInfo}{}}}
\subsection{Method \code{ShareArrayWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$ShareArrayWithHttpInfo(namespace, array, array.sharing, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateArrayMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateArrayMetadata}{}}}
\subsection{Method \code{UpdateArrayMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$UpdateArrayMetadata(namespace, array, array.metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateArrayMetadataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateArrayMetadataWithHttpInfo}{}}}
\subsection{Method \code{UpdateArrayMetadataWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$UpdateArrayMetadataWithHttpInfo(namespace, array, array.metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
