% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_metadata_entry.R
\docType{class}
\name{ArrayMetadataEntry}
\alias{ArrayMetadataEntry}
\title{ArrayMetadataEntry}
\format{
An \code{R6Class} generator object
}
\description{
ArrayMetadataEntry Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{key}}{character [optional]}

\item{\code{type}}{character [optional]}

\item{\code{value_num}}{integer [optional]}

\item{\code{value}}{list( integer ) [optional]}

\item{\code{del}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ArrayMetadataEntry$new()}}
\item \href{#method-toJSON}{\code{ArrayMetadataEntry$toJSON()}}
\item \href{#method-fromJSON}{\code{ArrayMetadataEntry$fromJSON()}}
\item \href{#method-toJSONString}{\code{ArrayMetadataEntry$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ArrayMetadataEntry$fromJSONString()}}
\item \href{#method-clone}{\code{ArrayMetadataEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$new(
  key = NULL,
  type = NULL,
  value_num = NULL,
  value = NULL,
  del = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$fromJSON(ArrayMetadataEntryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$fromJSONString(ArrayMetadataEntryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayMetadataEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
