% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_schema.R
\docType{class}
\name{ArraySchema}
\alias{ArraySchema}
\title{ArraySchema}
\format{
An \code{R6Class} generator object
}
\description{
ArraySchema Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{character [optional]}

\item{\code{version}}{list( integer )}

\item{\code{arrayType}}{\link{ArrayType}}

\item{\code{tileOrder}}{\link{Layout}}

\item{\code{cellOrder}}{\link{Layout}}

\item{\code{capacity}}{integer}

\item{\code{coordsFilterPipeline}}{\link{FilterPipeline}}

\item{\code{offsetFilterPipeline}}{\link{FilterPipeline}}

\item{\code{domain}}{\link{Domain}}

\item{\code{attributes}}{list( \link{Attribute} )}

\item{\code{allowsDuplicates}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ArraySchema$new()}}
\item \href{#method-toJSON}{\code{ArraySchema$toJSON()}}
\item \href{#method-fromJSON}{\code{ArraySchema$fromJSON()}}
\item \href{#method-toJSONString}{\code{ArraySchema$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ArraySchema$fromJSONString()}}
\item \href{#method-clone}{\code{ArraySchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$new(
  version,
  arrayType,
  tileOrder,
  cellOrder,
  capacity,
  coordsFilterPipeline,
  offsetFilterPipeline,
  domain,
  attributes,
  uri = NULL,
  allowsDuplicates = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$fromJSON(ArraySchemaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$fromJSONString(ArraySchemaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
