% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure TileDB Cloud}
\usage{
config()
}
\value{
A named vector with configuration values is returned.
}
\description{
Provide the setup configuration for the TileDB Cloud package.
}
\details{
It considers four different environment variables:
\code{TILEDB_REST_TOKEN}, \code{TILEDB_REST_HOST},
\code{TILEDB_REST_USERNAME}, and \code{TILEDB_REST_PASSWORD}.

It also reads a configuration file from a file
\code{.tiledb/httpshocloud.json}

To operate, \emph{either} an API token has to be provided and will
be used, \emph{or} the username and password combination will be
used to login with a new session.
}
