% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_array.R
\docType{class}
\name{DomainArray}
\alias{DomainArray}
\title{DomainArray}
\format{
An \code{R6Class} generator object
}
\description{
DomainArray Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{int8}}{list( integer ) [optional]}

\item{\code{uint8}}{list( integer ) [optional]}

\item{\code{int16}}{list( integer ) [optional]}

\item{\code{uint16}}{list( integer ) [optional]}

\item{\code{int32}}{list( integer ) [optional]}

\item{\code{uint32}}{list( integer ) [optional]}

\item{\code{int64}}{list( integer ) [optional]}

\item{\code{uint64}}{list( integer ) [optional]}

\item{\code{float32}}{list( numeric ) [optional]}

\item{\code{float64}}{list( numeric ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DomainArray$new()}}
\item \href{#method-toJSON}{\code{DomainArray$toJSON()}}
\item \href{#method-fromJSON}{\code{DomainArray$fromJSON()}}
\item \href{#method-toJSONString}{\code{DomainArray$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DomainArray$fromJSONString()}}
\item \href{#method-clone}{\code{DomainArray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$new(
  int8 = NULL,
  uint8 = NULL,
  int16 = NULL,
  uint16 = NULL,
  int32 = NULL,
  uint32 = NULL,
  int64 = NULL,
  uint64 = NULL,
  float32 = NULL,
  float64 = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$fromJSON(DomainArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$fromJSONString(DomainArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
