% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook_status.R
\docType{class}
\name{NotebookStatus}
\alias{NotebookStatus}
\title{NotebookStatus}
\format{
An \code{R6Class} generator object
}
\description{
NotebookStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{namespace}}{character [optional]}

\item{\code{uptime}}{integer [optional]}

\item{\code{cpu_usage}}{integer [optional]}

\item{\code{memory_usage}}{integer [optional]}

\item{\code{memory_limit}}{integer [optional]}

\item{\code{cpu_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NotebookStatus$new()}}
\item \href{#method-toJSON}{\code{NotebookStatus$toJSON()}}
\item \href{#method-fromJSON}{\code{NotebookStatus$fromJSON()}}
\item \href{#method-toJSONString}{\code{NotebookStatus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{NotebookStatus$fromJSONString()}}
\item \href{#method-clone}{\code{NotebookStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$new(
  namespace = NULL,
  uptime = NULL,
  cpu_usage = NULL,
  memory_usage = NULL,
  memory_limit = NULL,
  cpu_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$fromJSON(NotebookStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$fromJSONString(NotebookStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
