% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_user.R
\docType{class}
\name{OrganizationUser}
\alias{OrganizationUser}
\title{OrganizationUser}
\format{
An \code{R6Class} generator object
}
\description{
OrganizationUser Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{character [optional]}

\item{\code{organization_id}}{character [optional]}

\item{\code{username}}{character [optional]}

\item{\code{organization_name}}{character [optional]}

\item{\code{role}}{\link{OrganizationRoles} [optional]}

\item{\code{allowed_actions}}{list( \link{NamespaceActions} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrganizationUser$new()}}
\item \href{#method-toJSON}{\code{OrganizationUser$toJSON()}}
\item \href{#method-fromJSON}{\code{OrganizationUser$fromJSON()}}
\item \href{#method-toJSONString}{\code{OrganizationUser$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OrganizationUser$fromJSONString()}}
\item \href{#method-clone}{\code{OrganizationUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$new(
  user_id = NULL,
  organization_id = NULL,
  username = NULL,
  organization_name = NULL,
  role = NULL,
  allowed_actions = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$fromJSON(OrganizationUserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$fromJSONString(OrganizationUserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
