% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\docType{class}
\name{Pricing}
\alias{Pricing}
\title{Pricing}
\format{
An \code{R6Class} generator object
}
\description{
Pricing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{array_uuid}}{character [optional]}

\item{\code{pricing_name}}{character [optional]}

\item{\code{pricing_type}}{\link{PricingType} [optional]}

\item{\code{product_name}}{character [optional]}

\item{\code{product_statement_descriptor}}{character [optional]}

\item{\code{product_unit_label}}{\link{PricingUnitLabel} [optional]}

\item{\code{currency}}{\link{PricingCurrency} [optional]}

\item{\code{aggregate_usage}}{\link{PricingAggregateUsage} [optional]}

\item{\code{interval}}{\link{PricingInterval} [optional]}

\item{\code{divided_by}}{integer [optional]}

\item{\code{charge}}{numeric [optional]}

\item{\code{activated}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Pricing$new()}}
\item \href{#method-toJSON}{\code{Pricing$toJSON()}}
\item \href{#method-fromJSON}{\code{Pricing$fromJSON()}}
\item \href{#method-toJSONString}{\code{Pricing$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Pricing$fromJSONString()}}
\item \href{#method-clone}{\code{Pricing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$new(
  id = NULL,
  array_uuid = NULL,
  pricing_name = NULL,
  pricing_type = NULL,
  product_name = NULL,
  product_statement_descriptor = NULL,
  product_unit_label = NULL,
  currency = NULL,
  aggregate_usage = NULL,
  interval = NULL,
  divided_by = NULL,
  charge = NULL,
  activated = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$fromJSON(PricingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$fromJSONString(PricingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pricing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
