% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_api.R
\docType{class}
\name{StatsApi}
\alias{StatsApi}
\title{Stats operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Stats
}
\section{Methods}{

\describe{
\strong{ GetTiledbStats } \emph{  }
Fetch libtiledb stat

\itemize{
\item \emph{ @returnType } \link{InlineResponse200} \cr


\item status code : 200 | stats

\item return type : InlineResponse200 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  GetTiledbStats  ####################

library(tiledbcloud)

api.instance <- StatsApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetTiledbStats()


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StatsApi$new()}}
\item \href{#method-GetTiledbStats}{\code{StatsApi$GetTiledbStats()}}
\item \href{#method-GetTiledbStatsWithHttpInfo}{\code{StatsApi$GetTiledbStatsWithHttpInfo()}}
\item \href{#method-clone}{\code{StatsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTiledbStats"></a>}}
\if{latex}{\out{\hypertarget{method-GetTiledbStats}{}}}
\subsection{Method \code{GetTiledbStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$GetTiledbStats(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTiledbStatsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetTiledbStatsWithHttpInfo}{}}}
\subsection{Method \code{GetTiledbStatsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$GetTiledbStatsWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
