% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarray_partitioner.R
\docType{class}
\name{SubarrayPartitioner}
\alias{SubarrayPartitioner}
\title{SubarrayPartitioner}
\format{
An \code{R6Class} generator object
}
\description{
SubarrayPartitioner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subarray}}{\link{Subarray} [optional]}

\item{\code{budget}}{list( \link{AttributeBufferSize} ) [optional]}

\item{\code{current}}{\link{SubarrayPartitionerCurrent} [optional]}

\item{\code{state}}{\link{SubarrayPartitionerState} [optional]}

\item{\code{memoryBudget}}{integer [optional]}

\item{\code{memoryBudgetVar}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubarrayPartitioner$new()}}
\item \href{#method-toJSON}{\code{SubarrayPartitioner$toJSON()}}
\item \href{#method-fromJSON}{\code{SubarrayPartitioner$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubarrayPartitioner$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubarrayPartitioner$fromJSONString()}}
\item \href{#method-clone}{\code{SubarrayPartitioner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$new(
  subarray = NULL,
  budget = NULL,
  current = NULL,
  state = NULL,
  memoryBudget = NULL,
  memoryBudgetVar = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$fromJSON(SubarrayPartitionerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$fromJSONString(SubarrayPartitionerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitioner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
