% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf.R
\docType{class}
\name{UDF}
\alias{UDF}
\title{UDF}
\format{
An \code{R6Class} generator object
}
\description{
UDF Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{udf_info_name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{ranges}}{\link{UDFRanges} [optional]}

\item{\code{subarray}}{\link{UDFSubarray} [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{buffers}}{list( character ) [optional]}

\item{\code{result_format}}{\link{UDFResultType} [optional]}

\item{\code{task_name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UDF$new()}}
\item \href{#method-toJSON}{\code{UDF$toJSON()}}
\item \href{#method-fromJSON}{\code{UDF$fromJSON()}}
\item \href{#method-toJSONString}{\code{UDF$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UDF$fromJSONString()}}
\item \href{#method-clone}{\code{UDF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$new(
  udf_info_name = NULL,
  language = NULL,
  version = NULL,
  image_name = NULL,
  ranges = NULL,
  subarray = NULL,
  exec = NULL,
  exec_raw = NULL,
  buffers = NULL,
  result_format = NULL,
  task_name = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$fromJSON(UDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$fromJSONString(UDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
