% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_info_update.R
\docType{class}
\name{UDFInfoUpdate}
\alias{UDFInfoUpdate}
\title{UDFInfoUpdate}
\format{
An \code{R6Class} generator object
}
\description{
UDFInfoUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{type}}{\link{UDFType} [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{readme}}{character [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UDFInfoUpdate$new()}}
\item \href{#method-toJSON}{\code{UDFInfoUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{UDFInfoUpdate$fromJSON()}}
\item \href{#method-toJSONString}{\code{UDFInfoUpdate$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UDFInfoUpdate$fromJSONString()}}
\item \href{#method-clone}{\code{UDFInfoUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$new(
  name = NULL,
  language = NULL,
  version = NULL,
  image_name = NULL,
  type = NULL,
  exec = NULL,
  exec_raw = NULL,
  readme = NULL,
  license_id = NULL,
  license_text = NULL,
  tags = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$fromJSON(UDFInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$fromJSONString(UDFInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
