% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_info_version.R
\docType{class}
\name{UDFInfoVersion}
\alias{UDFInfoVersion}
\title{UDFInfoVersion}
\format{
An \code{R6Class} generator object
}
\description{
UDFInfoVersion Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{udf_image_uuid}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{default}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UDFInfoVersion$new()}}
\item \href{#method-toJSON}{\code{UDFInfoVersion$toJSON()}}
\item \href{#method-fromJSON}{\code{UDFInfoVersion$fromJSON()}}
\item \href{#method-toJSONString}{\code{UDFInfoVersion$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UDFInfoVersion$fromJSONString()}}
\item \href{#method-clone}{\code{UDFInfoVersion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$new(
  id = NULL,
  udf_image_uuid = NULL,
  name = NULL,
  version = NULL,
  image_name = NULL,
  exec = NULL,
  exec_raw = NULL,
  default = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$fromJSON(UDFInfoVersionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$fromJSONString(UDFInfoVersionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoVersion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
