% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\docType{class}
\name{User}
\alias{User}
\title{User}
\format{
An \code{R6Class} generator object
}
\description{
User Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{username}}{character}

\item{\code{password}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{is_valid_email}}{character [optional]}

\item{\code{stripe_connect}}{character [optional]}

\item{\code{company}}{character [optional]}

\item{\code{logo}}{character [optional]}

\item{\code{last_activity_date}}{character [optional]}

\item{\code{timezone}}{character [optional]}

\item{\code{organizations}}{list( \link{OrganizationUser} ) [optional]}

\item{\code{allowed_actions}}{list( \link{NamespaceActions} ) [optional]}

\item{\code{enabled_features}}{list( character ) [optional]}

\item{\code{unpaid_subscription}}{character [optional]}

\item{\code{notebook_settings}}{\link{NotebookSettings} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-toJSON}{\code{User$toJSON()}}
\item \href{#method-fromJSON}{\code{User$fromJSON()}}
\item \href{#method-toJSONString}{\code{User$toJSONString()}}
\item \href{#method-fromJSONString}{\code{User$fromJSONString()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(
  username,
  id = NULL,
  password = NULL,
  name = NULL,
  email = NULL,
  is_valid_email = NULL,
  stripe_connect = NULL,
  company = NULL,
  logo = NULL,
  last_activity_date = NULL,
  timezone = NULL,
  organizations = NULL,
  allowed_actions = NULL,
  enabled_features = NULL,
  unpaid_subscription = NULL,
  notebook_settings = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSON(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSONString(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
